rename RxWebSocket -> WalletLinkWebSocket
